v1.64.5

+++ app/src/main/java/com/idunnololz/summit/api/dto/BanPerson.kt
@@ -5,6 +5,6 @@ data class BanPerson(
   val ban: Boolean,
   val remove_data: Boolean? = null,
   val reason: String? = null,
-  val expires: Int? = null,
+  val expires: Long? = null,
   val auth: String,
 )
